refactor: remove utility wrappers
