--- src/utils/dependencyValidators.js
@@ -1,5 +1,0 @@
-export {
-  assertPresent,
-  assertFunction,
-  assertMethods,
-} from './dependencyUtils.js';
