--- src/utils/validationUtils.js
@@ -1,1 +1,0 @@
-export { validateDependency, validateDependencies } from './dependencyUtils.js';
